+++ src/UniversityEntranceQualification/UniversityEntranceQualification.mjs
@@ -1,7 +1,6 @@
 /** @typedef {import("./ChosenUniversityEntranceQualification.mjs").ChosenUniversityEntranceQualification} ChosenUniversityEntranceQualification */
-/** @typedef {import("./UniversityEntranceQualificationData.mjs").UniversityEntranceQualificationData} UniversityEntranceQualificationData */
 /** @typedef {import("./UniversityEntranceQualificationSelect.mjs").UniversityEntranceQualificationSelect} UniversityEntranceQualificationSelect */
 
 /**
- * @typedef {{data: UniversityEntranceQualificationData[][], "select-to-data": [string, number][], selects: UniversityEntranceQualificationSelect[], "select-index": number, values: ChosenUniversityEntranceQualification | null}} UniversityEntranceQualification
+ * @typedef {{data: {[key: string]: {[key: string]: *}}[], "select-to-data": [string, number][], selects: UniversityEntranceQualificationSelect[], "select-index": number, values: ChosenUniversityEntranceQualification | null}} UniversityEntranceQualification
  */